% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_org_alias.R
\name{clean_org_alias}
\alias{clean_org_alias}
\title{Clean Corporate Suffixes}
\usage{
clean_org_alias(x)
}
\arguments{
\item{x}{A character vector of names.}
}
\value{
A cleaned character vector.
}
\description{
Removes common legal suffixes and trailing punctuation from entity names.
}
