% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corporations_data}
\alias{corporations_data}
\title{corporations_data dataset}
\format{
A tibble with 5 columns:
\describe{
  \item{aliases}{aliases of a certain company (each one seperated by '|')}
  \item{cik}{Central Index Key of a corporation (numeric)}
  \item{FED_RSSD}{Research, Statistics, Supervision, and Discount identifier (numeric)}
  \item{ticker}{Stock ticker of a corporation}
  \item{naics}{North American Industry Classification System (NAICS) codes (numeric)}
}
}
\source{
Generated for the `corporations` package
}
\description{
A crosswalk lookup table of hundreds of thousands of corporations for regular expression matching and company searching
}
