% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_hearing_excerpt}
\alias{house_hearing_excerpt}
\title{house_hearing_excerpt dataset}
\format{
A tibble with 1 column:
\describe{
  \item{text}{A column wiht each row containing a sentence from the excerpt}
}
}
\source{
\url{https://www.congress.gov/119/chrg/CHRG-119hhrg62658/CHRG-119hhrg62658.pdf}
}
\description{
Sample text dataset used for demonstration of `corporations` and `search` mode extraction.
}
\details{
An excerpt of text converted to a tibble from a 119 Congress hearing titled "Embedded Threats:
Foreign Ownership, Hidden Hardware, and Licensing Failures in America’s Transportation Systems"
}
