% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corporations_data}
\alias{corporations_data}
\title{corporations_data dataset}
\format{
A tibble with ____ columns:
\describe{
  \item{column_1}{Description of column 1}
  \item{column_2}{Description of column 2}
}
}
\source{
\url{http://}
}
\usage{
corporations_data
}
\description{
A crosswalk lookup table of hundreds of thousands of corporations for regex matching and company searching
TODO: ONCE FINAL CROSSWALK IS DESIGNED, ADD DOCUMENTATION FOR THIS DATA
}
\keyword{datasets}
