% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_2025_coalition_and_contributors}
\alias{project_2025_coalition_and_contributors}
\title{project_2025_coalition_and_contributors dataset}
\format{
A tibble with ____ columns:
\describe{
  \item{column_1}{Description of column 1}
  \item{column_2}{Description of column 2}
}
}
\source{
\url{https://static.heritage.org/project2025/2025_MandateForLeadership_FULL.pdf}
}
\description{
Extracted corporations and individual contributors from Project 2025
}
