% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_fix.R
\name{report_fix}
\alias{report_fix}
\title{Report corporations matches}
\usage{
report_fix(results, corporations_data = NULL)
}
\arguments{
\item{results}{The dataframe returned from the extract() function.}

\item{corporations_data}{The built-in data for reference (optional here).}
}
\description{
Flag incorrect name matches or suggest missing aliases.
}
