% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_fix.R
\name{report_fix}
\alias{report_fix}
\title{Report corporations matches}
\usage{
report_fix(results)
}
\arguments{
\item{results}{The tibble returned from the extract() function.}
}
\description{
Flag incorrect name matches or suggest missing aliases.
}
